Refactor segmentation UI structure
